cast to float

+++ src/eco.py
@@ -238,7 +238,7 @@ def increase_skill_gain(ctx: invoke.Context, multiplier: float):
         skill_gain_multiplier = balance["SkillGainMultiplier"]
         print("Current Skill Gain Multiplier:", skill_gain_multiplier)
         balance["SkillGainMultiplier"] = round(
-            int(skill_gain_multiplier) * int(multiplier), 2
+            float(skill_gain_multiplier) * float(multiplier), 2
         )
 
     with open(file_path, "w", encoding="utf-8") as file:
